+++ etc/build_dir.sh
@@ -71,7 +71,7 @@ export PKG_CONFIG_PATH="$PWD/${BUILD_DIR}/install/lib/pkgconfig:${PKG_CONFIG_PAT
 export PYTHONPATH="$PWD/${BUILD_DIR}/install/lib/python${PYTHON_VERSION}:$PWD/${BUILD_DIR}/install/lib/python${PYTHON_VERSION}/site-packages:${PYTHONPATH}"
 export PYTHONEXE="$PWD/${BUILD_DIR}/install/bin/python${PYTHON_VERSION}"
 if [[ $KERNEL == *Msys* ]]; then
-    export PYTHONEXE="$PWD/${BUILD_DIR}/install/lib/python.exe"
+    export PYTHONEXE="$PWD/${BUILD_DIR}/install/bin/python.exe"
 fi
 
 if [[ $RUNME == 1 && $0 != *runme.sh* ]]; then